make data types more robust
